feat: 분석 페이지 기능 개선 및 UI 구성 요소 추가
