+++ src/entities/analysis/index.ts
@@ -1,0 +1,2 @@
+// ui
+export { AnalysisDisclaimer } from './ui/AnalysisDisclaimer';
